Handle content id, not project slug
