+++ app/routes.ts
@@ -9,10 +9,7 @@ export const Routes = {
   AdminCreateProject: `/admin/create-project`,
   AdminContent: (slug: string) => `/admin/content/${slug}`,
   AdminContentStatus: (slug: string) => `/admin/content/status/${slug}`,
-  AdminContentTitle: `/admin/content/title`,
-  AdminContentVideo: (slug: string) => `/admin/content/video/${slug}`,
-  AdminContentScheduler: (slug: string) => `/admin/content/scheduler/${slug}`,
-  AdminContentDetails: (slug: string) => `/admin/content/details/${slug}`,
+  AdminContentUpload: `/admin/content/upload`,
   Login: `/login`,
   Logout: `/logout`,
   Signup: `/signup`,
